chore: do not allow to delete tag in unse

+++ src/api/services/blog/Blog.Api/V1/Tags/Commands/DeleteTagCommand.cs
@@ -41,6 +41,13 @@ public class DeleteTagCommandHandler(
             throw new Exception($"Tag with id {command.Id} not found");
         }
 
+        if (tag.IsInUse())
+        {
+            logger.LogWarning("Tag with id {id} is in use", command.Id);
+
+            throw new Exception($"Tag with id {command.Id} is in use");
+        }
+
         repository.Remove(tag);
 
         await repository.UnitOfWork.SaveEntitiesAsync(cancellationToken);
